Fixed ICE-4884:     Updated version to 3.5b     Some fixes to distribution/bin/fixVersion.py

+++ cpp/src/IceGrid/Parser.cpp
@@ -2263,7 +2263,7 @@ Parser::getInput(char* buf, int& result, int maxSize)
             }
         }
         
-        result = line.length();
+        result = static_cast<int>(line.length());
         if(result > maxSize)
         {
             error("input line too long");
